Move SectorScene to a separate file
