+++ viewer/src/models/cad/CadModel.ts
@@ -2,7 +2,8 @@
  * Copyright 2020 Cognite AS
  */
 
-import { SectorScene, SectorModelTransformation } from './types';
+import { SectorModelTransformation } from './types';
+import { SectorScene } from './SectorScene';
 import { ModelDataRetriever } from '../../datasources/ModelDataRetriever';
 
 export interface CadModel {
