+++ viewer/src/views/threejs/cad/fromThreeCameraConfig.ts
@@ -7,7 +7,8 @@ import { vec3, mat4 } from 'gl-matrix';
 import { fromThreeVector3, fromThreeMatrix } from '../utilities';
 import { map } from 'rxjs/operators';
 import { OperatorFunction } from 'rxjs';
-import { SectorModelTransformation, SectorScene } from '../../../models/cad/types';
+import { SectorModelTransformation } from '../../../models/cad/types';
+import { SectorScene } from '../../../models/cad/SectorScene';
 import { DetermineSectorsByProximityInput } from '../../../models/cad/determineSectors';
 import { CadLoadingHints } from '../../../models/cad/CadLoadingHints';
